apps/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.1
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
cachetools>=5.3.2
python-json-logger>=2.0.7
fastapi-limiter>=0.1.5
prometheus-fastapi-instrumentator>=6.1.0
sqlalchemy>=2.0.23
alembic>=1.12.1
aiosqlite>=0.19.0
asyncpg>=0.29.0  # For PostgreSQL async support
greenlet>=3.0.0  # Required for SQLAlchemy async
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
email-validator>=2.1.0
aiofiles>=23.2.1

# Pipeline execution dependencies
python-slugify>=8.0.1
jinja2>=3.1.2
aiohttp>=3.8.6
jmespath>=1.0.1

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=19.13.0
asgi-lifespan>=2.1.0
